+++ opentelemetry-sdk/src/metrics/internal/mod.rs
@@ -1,4 +1,5 @@
 mod aggregate;
+mod exponential_histogram;
 mod histogram;
 mod last_value;
 mod sum;
@@ -7,6 +8,7 @@ use core::fmt;
 use std::ops::{Add, AddAssign, Sub};
 
 pub(crate) use aggregate::{AggregateBuilder, ComputeAggregation, Measure};
+pub(crate) use exponential_histogram::{EXPO_MAX_SCALE, EXPO_MIN_SCALE};
 
 pub(crate) trait Number<T>:
     Add<Output = T>
@@ -22,22 +24,49 @@ pub(crate) trait Number<T>:
     + Sync
     + 'static
 {
+    fn min() -> Self;
+    fn max() -> Self;
+
     fn into_float(self) -> f64;
 }
 
 impl Number<i64> for i64 {
+    fn min() -> Self {
+        i64::MIN
+    }
+
+    fn max() -> Self {
+        i64::MAX
+    }
+
     fn into_float(self) -> f64 {
         // May have precision loss at high values
         self as f64
     }
 }
 impl Number<u64> for u64 {
+    fn min() -> Self {
+        u64::MIN
+    }
+
+    fn max() -> Self {
+        u64::MAX
+    }
+
     fn into_float(self) -> f64 {
         // May have precision loss at high values
         self as f64
     }
 }
 impl Number<f64> for f64 {
+    fn min() -> Self {
+        f64::MIN
+    }
+
+    fn max() -> Self {
+        f64::MAX
+    }
+
     fn into_float(self) -> f64 {
         self
     }